+++ __tests__/rules/substituteAtRuleAliases.test.js
@@ -15,27 +15,27 @@ it('aliases @column-offset', () => {
   const output = `
     @media (min-width: 0){
       article{
-        margin-left: calc(8.3333333333% - 22.9166666667px)
+        margin-left: calc(8.333333333333334% - 22.916666666666668px)
       }
     }
     @media (min-width: 740px){
       article{
-        margin-left: calc(8.3333333333% - 32.0833333333px)
+        margin-left: calc(8.333333333333334% - 32.083333333333336px)
       }
     }
     @media (min-width: 1024px){
       article{
-        margin-left: calc(8.3333333333% - 45.8333333333px)
+        margin-left: calc(8.333333333333334% - 45.833333333333336px)
       }
     }
     @media (min-width: 1399px){
       article{
-        margin-left: calc(8.3333333333% - 45.8333333333px)
+        margin-left: calc(8.333333333333334% - 45.833333333333336px)
       }
     }
     @media (min-width: 1900px){
       article{
-        margin-left: calc(8.3333333333% - 55px)
+        margin-left: calc(8.333333333333334% - 54.99999999999999px)
       }
     }
   `
@@ -56,27 +56,27 @@ it('aliases @column-offset negative value', () => {
   const output = `
     @media (min-width: 0){
       article{
-        margin-left: calc(-8.3333333333% - 27.0833333333px)
+        margin-left: calc(-8.333333333333334% - 27.083333333333332px)
       }
     }
     @media (min-width: 740px){
       article{
-        margin-left: calc(-8.3333333333% - 37.9166666667px)
+        margin-left: calc(-8.333333333333334% - 37.916666666666664px)
       }
     }
     @media (min-width: 1024px){
       article{
-        margin-left: calc(-8.3333333333% - 54.1666666667px)
+        margin-left: calc(-8.333333333333334% - 54.166666666666664px)
       }
     }
     @media (min-width: 1399px){
       article{
-        margin-left: calc(-8.3333333333% - 54.1666666667px)
+        margin-left: calc(-8.333333333333334% - 54.166666666666664px)
       }
     }
     @media (min-width: 1900px){
       article{
-        margin-left: calc(-8.3333333333% - 65px)
+        margin-left: calc(-8.333333333333334% - 65px)
       }
     }
   `
@@ -97,9 +97,10 @@ it('aliases @column-offset for single bp', () => {
   const output = `
     @media (min-width: 0) and (max-width: 739px){
       article{
-        margin-left: calc(8.3333333333% - 22.9166666667px)
+        margin-left: calc(8.333333333333334% - 22.916666666666668px)
       }
     }
+    
   `
 
   return run(input).then(result => {