+++ __tests__/rules/substituteResponsiveAtRules.test.js
@@ -549,8 +549,8 @@ it('parses multiple breakpoints slashed', () => {
         position: relative;
         flex-grow: 0;
         flex-shrink: 0;
-        flex-basis: calc(33.3333333333% - 46.6666666667px);
-        max-width: calc(33.3333333333% - 46.6666666667px)
+        flex-basis: calc(33.333333333333336% - 46.66666666666667px);
+        max-width: calc(33.333333333333336% - 46.66666666666667px - 0.002vw)
       }
     }
     @media (min-width: 1200px) and (max-width: 1559px){
@@ -558,8 +558,8 @@ it('parses multiple breakpoints slashed', () => {
         position: relative;
         flex-grow: 0;
         flex-shrink: 0;
-        flex-basis: calc(33.3333333333% - 53.3333333333px);
-        max-width: calc(33.3333333333% - 53.3333333333px)
+        flex-basis: calc(33.333333333333336% - 53.333333333333336px);
+        max-width: calc(33.333333333333336% - 53.333333333333336px - 0.002vw)
       }
     }
     @media (min-width: 1560px) and (max-width: 1919px){
@@ -567,8 +567,8 @@ it('parses multiple breakpoints slashed', () => {
         position: relative;
         flex-grow: 0;
         flex-shrink: 0;
-        flex-basis: calc(33.3333333333% - 66.6666666667px);
-        max-width: calc(33.3333333333% - 66.6666666667px)
+        flex-basis: calc(33.333333333333336% - 66.66666666666667px);
+        max-width: calc(33.333333333333336% - 66.66666666666667px - 0.002vw)
       }
     }
     @media (min-width: 1920px){
@@ -576,8 +576,8 @@ it('parses multiple breakpoints slashed', () => {
         position: relative;
         flex-grow: 0;
         flex-shrink: 0;
-        flex-basis: calc(33.3333333333% - 80px);
-        max-width: calc(33.3333333333% - 80px)
+        flex-basis: calc(33.333333333333336% - 80px);
+        max-width: calc(33.333333333333336% - 80px - 0.002vw)
       }
     }
   `